Install modules from URL
Added option to the module import page that allows an URL to be
entered and install a module directly from internet and GitHub. When
using a project home URL in GitHub it will install the latest code in
the master branch. In case that's not what you want please use the URL
pointing to the zip file of the desired release.

Automatically installing the latest release instead of the latest in
the master branch could be implemented with some more work.