Fix crmdetailviewurl in Workflows email template.
Now when we have to replate the module name to create the url to detailview record, we have to check if modulename is Events to replace to Calendar.
But how now we have Calendar4You, aslo we have to check Calendar4You to replace the url with te correct for this module.