Fix steps migration from 5.2.1 to 5.4
Now we don't use the zip packages to install the mandatory and optional modules.
We install all the modules importing the manifest.xml.
For this we have had to delete the functions call to package zip for not obtain an error
and just call to importManifest function.

Also we have found an error installing the default workflows in the migration, but
now is fixed.