Fix report problem with special field DISCOUNT
This field is a special field to get values from product/service lines in reports.
We had a problem with the columna name save in report and the spect value in code.
Now we have replaced the spect value discount_amount to discount. This fix the problem
that this columns was not show in the reports result.

An other problem was, that you can filter by this field if you applied a discount_percent,
the query always look into the discount_amount column in vtiger_inventoryproductrel.
Now this column and discount_percent are checked to a correct filter.