Optionally show my websites in iframe or another window
Based on response header of a given page, define if it can be shown in iframe or not

Https pages will allways show in another window due to certificate issues.

If default page can not be shown in iframe then the select will be filled
with blank option at the beginning.